Refactoring
- Removing entity references
- Remove two dead classes
