--- app/AllowedValue/ResourceItem.php
@@ -1,240 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace App\AllowedValue;
-
-use App\ItemType\ItemType as Entity;
-use App\Models\Category;
-use App\Models\EntityLimits;
-use App\Models\Subcategory;
-use App\Request\Hash;
-
-class ResourceItem
-{
-    private Hash $hash;
-
-    private Entity $entity;
-
-    private EntityLimits $model;
-
-    public function __construct(Entity $entity)
-    {
-        $this->hash = new Hash();
-
-        $this->entity = $entity;
-
-        $this->model = new EntityLimits();
-    }
-
-    /**
-     * @return array
-     */
-    public function allowedValues(
-        int $resource_type_id,
-        int $resource_id,
-        array $viewable_resource_types,
-        array $available_parameters,
-        array $defined_parameters,
-        bool $include_currencies = false
-    ): array
-    {
-        $years = [];
-        if (array_key_exists('year', $available_parameters)) {
-            $years = $this->allowedValuesForYear(
-                $resource_type_id,
-                $resource_id
-            );
-        }
-
-        $months = [];
-        if (array_key_exists('month', $available_parameters)) {
-            $months = $this->allowedValuesForMonth();
-        }
-
-        $categories = [];
-        if (array_key_exists('category', $available_parameters)) {
-            $categories = $this->allowedValuesForCategory(
-                $resource_type_id,
-                $viewable_resource_types
-            );
-        }
-
-        $subcategories = [];
-
-        if (
-            array_key_exists('category', $available_parameters) === true &&
-            array_key_exists('subcategory', $available_parameters) === true &&
-            array_key_exists('category', $defined_parameters) === true &&
-            $defined_parameters['category'] !== null
-        ) {
-            $subcategories = $this->allowedValuesForSubcategory(
-                $resource_type_id,
-                $defined_parameters['category']
-            );
-        }
-
-        $currencies = [];
-
-        if ($include_currencies === true) {
-            $currencies = $this->allowedValuesForCurrency();
-        }
-
-        return array_merge(
-            $years,
-            $months,
-            $categories,
-            $subcategories,
-            $currencies
-        );
-    }
-
-    protected function allowedValuesForCategory(
-        int $resource_type_id,
-        array $viewable_resource_types
-    ): array
-    {
-        $parameters = ['category' => ['allowed_values' => []]];
-
-        $categories = (new Category())->paginatedCollection(
-            $resource_type_id,
-            $viewable_resource_types,
-            0,
-            100
-        );
-
-        foreach ($categories as $category) {
-            $category_id = $this->hash->encode('category', $category['category_id']);
-
-            $parameters['category']['allowed_values'][$category_id] = [
-                'value' => $category_id,
-                'name' => $category['category_name'],
-                'description' => trans('item-type-' . $this->entity->type() .
-                        '/allowed-values.description-prefix-category') .
-                    $category['category_name'] .
-                    trans('item-type-' . $this->entity->type() .
-                        '/allowed-values.description-suffix-category')
-            ];
-        }
-
-        return $parameters;
-    }
-
-    /**
-     * @return array
-     */
-    protected function allowedValuesForMonth(): array
-    {
-        $parameters = ['month' => ['allowed_values' => []]];
-
-        for ($i=1; $i < 13; $i++) {
-            $parameters['month']['allowed_values'][$i] = [
-                'value' => $i,
-                'name' => date("F", mktime(0, 0, 0, $i, 10)),
-                'description' => trans('item-type-' . $this->entity->type() .
-                        '/allowed-values.description-prefix-month') .
-                    date("F", mktime(0, 0, 0, $i, 1))
-            ];
-        }
-
-        return $parameters;
-    }
-
-    protected function allowedValuesForCurrency(): array
-    {
-        $parameters = ['currency_id' => ['allowed_values' => []]];
-
-        $currencies = (new \App\Models\Currency())->minimisedCollection();
-
-        foreach ($currencies as $currency) {
-            $id = $this->hash->encode('currency', $currency['currency_id']);
-
-            if ($id === false) {
-                \App\Response\Responses::unableToDecode();
-            }
-
-            $parameters['currency_id']['allowed_values'][$id] = [
-                'value' => $id,
-                'name' => $currency['currency_name'],
-                'description' => $currency['currency_name']
-            ];
-        }
-
-        return $parameters;
-    }
-
-    /**
-     * @return array
-     */
-    protected function allowedValuesForSubcategory(
-        int $resource_type_id,
-        int $category_id
-    ): array
-    {
-        $parameters = ['subcategory' => ['allowed_values' => []]];
-
-        $subcategories = (new Subcategory())->paginatedCollection(
-            $resource_type_id,
-            $category_id
-        );
-
-        array_map(
-            function($subcategory) use (&$parameters) {
-                $subcategory_id = $this->hash->encode('subcategory', $subcategory['subcategory_id']);
-                $parameters['subcategory']['allowed_values'][$subcategory_id] = [
-                    'value' => $subcategory_id,
-                    'name' => $subcategory['subcategory_name'],
-                    'description' => trans('item-type-' . $this->entity->type() . '/allowed-values.description-prefix-subcategory') .
-                        $subcategory['subcategory_name'] . trans('item-type-' . $this->entity->type() . '/allowed-values.description-suffix-subcategory')
-                ];
-            },
-            $subcategories
-        );
-
-        return $parameters;
-    }
-
-    protected function allowedValuesForYear(
-        int $resource_type_id,
-        int $resource_id
-    ): array
-    {
-        $parameters = ['year' => ['allowed_values' => []]];
-
-        $min_year = null;
-        $max_year = null;
-
-        $item_type = \App\ItemType\Entity::itemType($resource_type_id);
-        switch($item_type) {
-            case 'allocated-expense':
-                $min_year = $this->model->minimumYearByResourceTypeAndResource(
-                    $resource_type_id,
-                    $resource_id,
-                    'item_type_allocated_expense',
-                    'effective_date'
-                );
-                $max_year = $this->model->maximumYearByResourceTypeAndResource(
-                    $resource_type_id,
-                    $resource_id,
-                    'item_type_allocated_expense',
-                    'effective_date'
-                );
-                break;
-            default:
-                // Do nothing
-                break;
-        }
-
-        if ($min_year !== null && $max_year !== null) {
-            for ($i = $min_year; $i <= $max_year; $i++) {
-                $parameters['year']['allowed_values'][$i] = [
-                    'value' => $i,
-                    'name' => $i,
-                    'description' => trans('item-type-' . $this->entity->type() .
-                            '/allowed-values.description-prefix-year') . $i
-                ];
-            }
-        }
-
-        return $parameters;
-    }
-}
