+++ app/ItemType/Game/AllowedValue/ResourceTypeItem.php
@@ -58,11 +58,9 @@ class ResourceTypeItem extends ResourceTypeAllowedValue
                 $allowed_values[$winner_id] = [
                     'value' => $winner_id,
                     'name' => $winner['category_name'],
-                    'description' => trans('resource-type-item-type-' . $this->entity->type() .
-                            '/allowed-values.description-prefix-winner_id') .
+                    'description' => trans('resource-type-item-type-game/allowed-values.description-prefix-winner_id') .
                         $winner['category_name'] .
-                        trans('resource-type-item-type-' . $this->entity->type() .
-                            '/allowed-values.description-suffix-winner_id')
+                        trans('resource-type-item-type-game/allowed-values.description-suffix-winner_id')
                 ];
             }
